refactor: 요구사항 확인 및 패키지 정리
요구사항 확인
패키지 정리
