+++ src/main/java/racingcar/domain/car/accelpower/AccelPowerStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.accelpower;
 
 public interface AccelPowerStrategy {
     boolean isMove();
