+++ src/main/java/racingcar/domain/car/accelpower/RandomAccelPowerStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.accelpower;
 
 import camp.nextstep.edu.missionutils.Randoms;
