+++ src/test/java/racingcar/domain/car/CarNamesTest.java
@@ -1,9 +1,10 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.CarNames;
 
 public class CarNamesTest {
     @Test
