+++ src/test/java/racingcar/domain/car/PositionTest.java
@@ -1,9 +1,10 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Position;
 
 public class PositionTest {
