+++ src/test/java/racingcar/domain/racetrack/LapTest.java
@@ -1,10 +1,11 @@
-package racingcar.domain;
+package racingcar.domain.racetrack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.racetrack.Lap;
 import racingcar.message.ERRORMessage;
 
 public class LapTest {
