+++ src/test/java/racingcar/domain/racetrack/RaceTrackTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.racetrack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.CarNames;
+import racingcar.domain.racetrack.Lap;
+import racingcar.domain.racetrack.RaceTrack;
 
 public class RaceTrackTest {
